pyndora/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
idna==3.3
mnemonic==0.20
pycparser==2.21
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
